FPS40, min_crossing->0.01

+++ src/LineSeg.cpp
@@ -55,12 +55,12 @@ LineSeg(float xs, float ys, float xe, float ye)
 //   DOES NOT tell you whether the edge is "entering" or "leaving".
 //   But you can use tests like Edge::Point_Side() to figure that out.
 //======================================================================
-float LineSeg::
+double LineSeg::
 Cross_Param(LineSeg e)
 //======================================================================
 {
-	float   dx1, dy1, dx2, dy2;
-	float   denom, s;
+	double   dx1, dy1, dx2, dy2;
+	double   denom, s;
 
 	// This computation comes from writing each segment in parametric form,
 	// and solving a simulataneous equation to determine the parameter
